+++ src/main/java/org/folio/persist/ItemRepository.java
@@ -7,12 +7,15 @@ import java.util.List;
 import java.util.Map;
 
 import io.vertx.core.AsyncResult;
+import org.folio.cql2pgjson.CQL2PgJSON;
 import org.folio.rest.jaxrs.model.Item;
 import org.folio.rest.persist.Criteria.Criteria;
 import org.folio.rest.persist.Criteria.Criterion;
-
+import org.folio.rest.persist.cql.CQLWrapper;
 import io.vertx.core.Context;
 import io.vertx.core.Future;
+import io.vertx.sqlclient.Row;
+import io.vertx.sqlclient.RowSet;
 import org.folio.rest.persist.SQLConnection;
 
 public class ItemRepository extends AbstractRepository<Item> {
@@ -26,4 +29,21 @@ public class ItemRepository extends AbstractRepository<Item> {
 
     return get(connection, criterion);
   }
+
+  /**
+   * Delete by CQL. For each deleted record return a {@link Row} with the instance id String
+   * and with the item's jsonb String.
+   */
+  public Future<RowSet<Row>> delete(String cql) {
+    try {
+      CQLWrapper cqlWrapper = new CQLWrapper(new CQL2PgJSON(tableName + ".jsonb"), cql, -1, -1);
+      String sql = "DELETE FROM " + postgresClientFuturized.getFullTableName(tableName)
+          + " " + cqlWrapper.getWhereClause()
+          + " RETURNING (SELECT instanceId::text FROM holdings_record WHERE id = holdingsRecordId), jsonb::text";
+      return postgresClient.execute(sql);
+    } catch (Exception e) {
+      return Future.failedFuture(e);
+    }
+  }
+
 }