+++ src/main/java/org/folio/rest/impl/InstanceStorageAPI.java
@@ -35,9 +35,11 @@ import org.folio.rest.persist.PgExceptionUtil;
 import org.folio.rest.persist.PgUtil;
 import org.folio.rest.persist.PostgresClient;
 import org.folio.rest.persist.cql.CQLWrapper;
+import org.folio.rest.support.EndpointFailureHandler;
 import org.folio.rest.tools.messages.MessageConsts;
 import org.folio.rest.tools.messages.Messages;
 import org.folio.rest.tools.utils.TenantTool;
+import org.folio.services.holding.HoldingsService;
 import org.folio.services.instance.InstanceService;
 
 public class InstanceStorageAPI implements InstanceStorage {
@@ -112,16 +114,14 @@ public class InstanceStorageAPI implements InstanceStorage {
   @Validate
   @Override
   public void deleteInstanceStorageInstances(
-    String lang,
+    String query,
     RoutingContext routingContext, Map<String, String> okapiHeaders,
     Handler<AsyncResult<Response>> asyncResultHandler,
     Context vertxContext) {
 
-    new InstanceService(vertxContext, okapiHeaders)
-      .deleteAllInstances()
-      .onSuccess(notUsed -> asyncResultHandler.handle(
-        succeededFuture(DeleteInstanceStorageInstancesResponse.respond204())))
-      .onFailure(handleFailure(asyncResultHandler));
+    new InstanceService(vertxContext, okapiHeaders).deleteInstances(query)
+    .otherwise(e -> EndpointFailureHandler.failureResponse(e))
+    .onComplete(asyncResultHandler);
   }
 
   @Validate