+++ src/main/java/org/folio/services/domainevent/ItemDomainEventPublisher.java
@@ -52,6 +52,11 @@ public class ItemDomainEventPublisher extends AbstractDomainEventPublisher<Item,
       .compose(domainEventService::publishRecordsUpdated);
   }
 
+  public void publishRemoved(String instanceId, String itemRaw) {
+    String instanceIdAndItemRaw = "{\"instanceId\":\"" + instanceId + "\"," + itemRaw.substring(1);
+    domainEventService.publishRecordRemoved(instanceId, instanceIdAndItemRaw);
+  }
+
   @Override
   protected String getId(Item record) {
     return record.getId();