+++ src/test/java/org/folio/rest/api/InstanceDomainEventTest.java
@@ -84,6 +84,6 @@ public class InstanceDomainEventTest extends TestBaseWithInventoryUtil {
 
     final Response removeResponse = instancesClient.attemptDeleteAll();
 
-    assertThat(removeResponse.getStatusCode(), is(500));
+    assertThat(removeResponse.getStatusCode(), is(400));
     assertNoRemoveEvent(instance.getId().toString());
   }}