+++ src/test/java/org/folio/rest/api/InstanceStorageTest.java
@@ -441,6 +441,32 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   @Test
+  public void canDeleteInstancesByCql() throws Exception {
+
+    var id5 = UUID.randomUUID();
+    var instance1 = createInstance(nod(UUID.randomUUID()).put("hrid", "1234")).getJson();
+    var instance2 = createInstance(nod(UUID.randomUUID()).put("hrid", "2123")).getJson();
+    var instance3 = createInstance(nod(UUID.randomUUID()).put("hrid", "12")).getJson();
+    var instance4 = createInstance(nod(UUID.randomUUID()).put("hrid", "345 12")).getJson();
+    var instance5 = createInstance(nod(id5).put("hrid", "123")).getJson();
+    put(id5, marcJson);
+    instance5 = getById(id5).getJson();
+
+    var response = client.delete(instancesStorageUrl("?query=hrid==12*"), TENANT_ID).get(5, SECONDS);
+
+    assertThat(response.getStatusCode(), is(HttpURLConnection.HTTP_NO_CONTENT));
+    assertExists(instance2);
+    assertExists(instance4);
+    assertNotExists(instance1);
+    assertNotExists(instance3);
+    assertNotExists(instance5);
+    getMarcJsonNotFound(id5);
+    assertRemoveEventForInstance(instance1);
+    assertRemoveEventForInstance(instance3);
+    assertRemoveEventForInstance(instance5);
+  }
+
+  @Test
   public void canGetInstanceById()
     throws MalformedURLException,
     InterruptedException,
@@ -1146,7 +1172,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
   }
 
   /**
-   * Create the 5 example instances and run a get request using the provided cql query.
+   * Create the 5 example instances.
    */
   private void create5instances() {
     try {
@@ -1560,7 +1586,7 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
 
     CompletableFuture<Response> allDeleted = new CompletableFuture<>();
 
-    client.delete(instancesStorageUrl(""), TENANT_ID,
+    client.delete(instancesStorageUrl("?query=cql.allRecords=1"), TENANT_ID,
       ResponseHandler.empty(allDeleted));
 
     Response deleteResponse = allDeleted.get(5, SECONDS);
@@ -2913,6 +2939,10 @@ public class InstanceStorageTest extends TestBaseWithInventoryUtil {
     Response response = getById(expectedInstance.getString("id"));
     assertThat(response, statusCodeIs(HttpStatus.HTTP_OK));
     assertThat(response.getBody(), containsString(expectedInstance.getString("title")));
+  }
+
+  private void assertNotExists(JsonObject instanceToGet) {
+    assertGetNotFound(instancesStorageUrl("/" + instanceToGet.getString("id")));
   }
 
   public static JsonObject smallAngryPlanet(UUID id) {